MAINT: fix more compiler warnings

+++ pywt/_extensions/_dwt.pxd
@@ -1,3 +1,4 @@
 from ._pywt cimport Wavelet, data_t
 
-cpdef upcoef(bint do_rec_a, data_t[::1] coeffs, Wavelet wavelet, int level, int take)
+cpdef upcoef(bint do_rec_a, data_t[::1] coeffs, Wavelet wavelet, int level,
+             size_t take)
